added dietary requirements to generate meal plan

+++ Backend/aspnet-core/src/LifeQuest.Application/Services/Health/Ingredient/Dtos/IngredientDto.cs
@@ -10,10 +10,10 @@ namespace LifeQuest.Services.Health.Ingredient.Dtos
     {
         public Guid Id { get; set; }
         public string Name { get; set; }
-        public int ServingSize { get; set; }
-        public int Calories { get; set; }
-        public int Protein { get; set; }
-        public int Carbohhydrates { get; set; }
-        public int Fats { get; set; }
+        public double ServingSize { get; set; }
+        public double Calories { get; set; }
+        public double Protein { get; set; }
+        public double Carbohhydrates { get; set; }
+        public double Fats { get; set; }
     }
 }
